requirements_tests.txt
numpyro>=0.9.1
jax>=0.3.1
matplotlib>=3.4.3
bokeh>=2.4.2
arviz>=0.12.0